move currency to separate package
